Main->Niko

+++ Zadanie 1/Task/inc/Task.hpp
@@ -3,26 +3,26 @@ class Task
 {
 private:
     /* private data */
-    double m_pj; // execution time
-    double m_rj; // availability time
-    double m_qj; // cooling time
+    int m_pj; // execution time
+    int m_rj; // availability time
+    int m_qj; // cooling time
     int m_taskId;
 
 public:
     /* public methods */
     Task();
-    Task(double pj, double rj, double qj, int id);
+    Task(int pj, int rj, int qj, int id);
     Task(const Task& task);
     ~Task();
     Task &operator=(const Task &task) noexcept;
 
-    double GetPj() const {return m_pj;};
-    double GetRj() const {return m_rj;};
-    double GetQj() const {return m_qj;};
+    int GetPj() const {return m_pj;};
+    int GetRj() const {return m_rj;};
+    int GetQj() const {return m_qj;};
     int GetTaskId() const {return m_taskId;};
 
-    void SetPj(double pj){ m_pj = pj;};
-    void SetRj(double rj){ m_rj = rj;};
-    void SetQj(double qj){ m_qj = qj;};
-    void SetTaskId(double id){ m_taskId = id;};
+    void SetPj(int pj){ m_pj = pj;};
+    void SetRj(int rj){ m_rj = rj;};
+    void SetQj(int qj){ m_qj = qj;};
+    void SetTaskId(int id){ m_taskId = id;};
 };
